Extract mentions hovercards into module
Move hovercard implementation out of src/modules/mentions/index.js into a new src/modules/mentions/hovercards.js. This removes hovercard-specific DOM/root logic from the main mentions entry, imports initMentionsHovercards from the new module, and lazy-loads hovercards in the audit-log viewer with a dedicated webpack chunk (mentions-hovercards). The change keeps hovercard behavior isolated so consumers (e.g. audit-log) can load hovercards without pulling editor/autocomplete code, and preserves caching, event delegation, and auth checks for showing hovercards.
